IE 11 scrollbar blur() bug
This fix is to address the dropdown disappearing in IE 11 when the user attempts to click the scroll bar.  When the user clicks the scrollbar, the blur event was being triggered and causing the dropdown to close.  The fix determines whether the user is still hovering over the list and prevents the dropdown from closing if so.